% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{ber_mean}
\alias{ber_mean}
\title{Find the mean of a Bernoulli distribution}
\usage{
ber_mean(p)
}
\arguments{
\item{p}{Probability of success}
}
\value{
Mean of the Bernoulli distribution
}
\description{
Find the mean of a Bernoulli distribution
}
