% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{ber_sd}
\alias{ber_sd}
\title{Find the standard deviation of a Bernoulli distribution}
\usage{
ber_sd(p)
}
\arguments{
\item{p}{Probability of success}
}
\value{
Standard deviation of the Bernoulli distribution
}
\description{
Find the standard deviation of a Bernoulli distribution
}
