% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{ber_var}
\alias{ber_var}
\title{Find the variance of a Bernoulli distribution}
\usage{
ber_var(p)
}
\arguments{
\item{p}{Probability of success}
}
\value{
Variance of the Bernoulli distribution
}
\description{
Find the variance of a Bernoulli distribution
}
