% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{class_mean}
\alias{class_mean}
\title{Find the mean of a class interval}
\usage{
class_mean(low, high, freq)
}
\arguments{
\item{low}{Lower bound of the class interval}

\item{high}{Upper bound of the class interval}

\item{freq}{Frequency of the class interval}
}
\value{
Mean of the class interval
}
\description{
Find the mean of a class interval
}
