% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{class_sd}
\alias{class_sd}
\title{Find the standard deviation of a class interval}
\usage{
class_sd(low, high, freq, pop = FALSE)
}
\arguments{
\item{low}{Lower bound of the class interval}

\item{high}{Upper bound of the class interval}

\item{freq}{Frequency of the class interval}

\item{pop}{Whether to use population or sample standard deviation}
}
\value{
Standard deviation of the class interval
}
\description{
Find the standard deviation of a class interval
}
