% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{class_var}
\alias{class_var}
\title{Find the variance of a class interval}
\usage{
class_var(low, high, freq, pop = FALSE)
}
\arguments{
\item{low}{Lower bound of the class interval}

\item{high}{Upper bound of the class interval}

\item{freq}{Frequency of the class interval}

\item{pop}{Whether to use population or sample variance}
}
\value{
Variance of the class interval
}
\description{
Find the variance of a class interval
}
