% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combos.R
\name{comb}
\alias{comb}
\title{Calculate combinations, N choose R / N c R}
\usage{
comb(n, k)
}
\arguments{
\item{n}{The total number of items}

\item{r}{The number of items to choose}
}
\value{
The number of combinations
}
\description{
Calculate combinations, N choose R / N c R
}
