% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{discrete_mean}
\alias{discrete_mean}
\title{Find the mean of a discrete distribution}
\usage{
discrete_mean(val, freq)
}
\arguments{
\item{val}{Values of the discrete distribution}

\item{freq}{Frequency of the discrete distribution}
}
\value{
Mean of the discrete distribution
}
\description{
Find the mean of a discrete distribution
}
