% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{discrete_sd}
\alias{discrete_sd}
\title{Find the standard deviation of a discrete distribution}
\usage{
discrete_sd(val, freq)
}
\arguments{
\item{val}{Values of the discrete distribution}

\item{freq}{Frequency of the discrete distribution}
}
\value{
Standard deviation of the discrete distribution
}
\description{
Find the standard deviation of a discrete distribution
}
