% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{discrete_var}
\alias{discrete_var}
\title{Find the variance of a discrete distribution}
\usage{
discrete_var(val, freq)
}
\arguments{
\item{val}{Values of the discrete distribution}

\item{freq}{Frequency of the discrete distribution}
}
\value{
Variance of the discrete distribution
}
\description{
Find the variance of a discrete distribution
}
