% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{drvvar}
\alias{drvvar}
\title{Calculate variance for a discreet random variable given a PDF}
\usage{
drvvar(x, prob)
}
\arguments{
\item{x}{The set of all possilbe values}

\item{r}{The set of probabilities for all expected values}
}
\value{
The variance for the sets
}
\description{
NOTE - Length of x must equal lenght of prob, and prob must sum to 1
}
