% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{imp}
\alias{imp}
\title{Prints the imperical rule}
\usage{
imp()
}
\description{
The imperical rule states that for a normal distribution:
\itemize{
\item 68.26\% of the data falls within 1 standard deviation of the mean
\item 95.44\% of the data falls within 2 standard deviations of the mean
\item 99.74\% of the data falls within 3 standard deviations of the mean
}
}
