% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combos.R, R/testing.R
\name{perm}
\alias{perm}
\title{Calculate permutations, N choose R / N p R}
\usage{
perm(n, r)

perm(n, r)
}
\arguments{
\item{n}{The total number of items}

\item{r}{The number of items to choose}
}
\value{
The number of permutations

The number of permutations
}
\description{
Calculate permutations, N choose R / N p R

Calculate permutations, N choose R / N p R
}
