% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{pop_cv}
\alias{pop_cv}
\title{Find population coefficient of variation}
\usage{
pop_cv(s, pop = TRUE)
}
\arguments{
\item{Set}{of numbers}

\item{Whether}{to use population or sample coefficient of variation}
}
\value{
Population coefficient of variation
}
\description{
Find population coefficient of variation
}
