% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{pop_sd}
\alias{pop_sd}
\title{Find population standard deviation}
\usage{
pop_sd(s, pop = TRUE)
}
\arguments{
\item{Set}{of numbers}

\item{Whether}{to use population or sample standard deviation}
}
\value{
Population standard deviation
}
\description{
Find population standard deviation
}
