% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{pop_var}
\alias{pop_var}
\title{Find population variance}
\usage{
pop_var(s, pop = TRUE)
}
\arguments{
\item{Set}{of numbers}

\item{Whether}{to use population or sample variance}
}
\value{
Population variance
}
\description{
Find population variance
}
