% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{unif_mean}
\alias{unif_mean}
\title{Find the mean of a binomial distribution}
\usage{
unif_mean(min, max)
}
\arguments{
\item{n}{Number of trials}

\item{p}{Probability of success}
}
\value{
Mean of the binomial distribution
}
\description{
Find the mean of a binomial distribution
}
\examples{
bin_mean(38, 45)
}
