% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{unif_sd}
\alias{unif_sd}
\title{Find the standard deviation of a uniform distribution}
\usage{
unif_sd(min, max)
}
\arguments{
\item{min}{Minimum value of the uniform distribution}

\item{max}{Maximum value of the uniform distribution}
}
\value{
Standard deviation of the uniform distribution
}
\description{
Find the standard deviation of a uniform distribution
}
\examples{
unif_sd(1, 10)
unif_sd(5, 15)
}
