% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{unif_var}
\alias{unif_var}
\title{Find the variance of a uniform distribution}
\usage{
unif_var(min, max)
}
\arguments{
\item{min}{Minimum value of the uniform distribution}

\item{max}{Maximum value of the uniform distribution}
}
\value{
Variance of the uniform distribution
}
\description{
Find the variance of a uniform distribution
}
\examples{
unif_var(1, 10)
unif_var(5, 15)
}
