% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{uniform_mean}
\alias{uniform_mean}
\title{Find the mean of a uniform distribution}
\usage{
uniform_mean(x)
}
\arguments{
\item{x}{set of values}
}
\value{
Mean of the uniform distribution
}
\description{
Find the mean of a uniform distribution
}
