% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{uniform_sd}
\alias{uniform_sd}
\title{Find the standard deviation of a uniform distribution}
\usage{
uniform_sd(x)
}
\arguments{
\item{x}{set of values}
}
\value{
Standard deviation of the uniform distribution
}
\description{
Find the standard deviation of a uniform distribution
}
