% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drv.R
\name{uniform_var}
\alias{uniform_var}
\title{Find the variance of a uniform distribution}
\usage{
uniform_var(x)
}
\arguments{
\item{x}{set of values}
}
\value{
Variance of the uniform distribution
}
\description{
Find the variance of a uniform distribution
}
